requirements.txt
Django
djangorestframework
python-dotenv
djangorestframework-simplejwt
drf-yasg
gunicorn
psycopg2-binary
pillow